# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SN Auto Parts API
  description: |
    Feature-config driven e-commerce API for auto parts.
    
    ## Authentication
    All authenticated endpoints require a session token passed via:
    - Cookie: `auth-token=<token>`
    - Header: `Authorization: Bearer <token>`
    
    ## Authorization
    APIs are generic - authorization is determined by the authenticated user's role 
    and feature configuration. The same endpoint may return different data or 
    permissions based on the user's role.
    
    ## Feature-Based Access
    Each endpoint requires specific features to be enabled in the user's role configuration.
    If a user lacks the required feature, they receive a 403 Forbidden response.
  version: 1.0.0
  contact:
    name: SN Auto Parts
    email: support@snautoparts.com
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.snautoparts.com/api/v1
    description: Production server
...
paths:
...
  /profile:
    get:
      tags: [Profile]
      summary: Get own profile
      operationId: getProfile
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Profile data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        $ref: '#/components/schemas/User'
    patch:
      tags: [Profile]
      summary: Update own profile
      operationId: updateProfile
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        $ref: '#/components/schemas/User'
  /profile/addresses:
    get:
      tags: [Profile]
      summary: List own addresses
      operationId: listAddresses
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Addresses list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Address'
    post:
      tags: [Profile]
      summary: Create address
      operationId: createAddress
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAddressRequest'
      responses:
        '201':
          description: Address created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Address'
  /profile/addresses/{id}:
    patch:
      tags: [Profile]
      summary: Update address
      operationId: updateAddress
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAddressRequest'
      responses:
        '200':
          description: Address updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Address'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Profile]
      summary: Delete address
      operationId: deleteAddress
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # SETTINGS ENDPOINTS
  # ============================================
...
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Session token (also accepted as auth-token cookie)
...
  responses:
    Success:
      description: Operation successful
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    DataWrapper:
      type: object
      properties:
        data: {}
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
        details:
          type: object

    # Auth Schemas
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [CUSTOMER, MANAGER, ADMIN]
        image:
          type: string
        createdAt:
          type: string
          format: date-time
...
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string

    # Feature Config
...
    Address:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        label:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        street:
          type: string
        apartment:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string
        phone:
          type: string
        isDefault:
          type: boolean
    CreateAddressRequest:
      type: object
      required: [firstName, lastName, street, city, state, zipCode]
      properties:
        label:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        street:
          type: string
        apartment:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string
          default: US
        phone:
          type: string
        isDefault:
          type: boolean
    UpdateAddressRequest:
      type: object
      properties:
        label:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        street:
          type: string
        apartment:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string
        phone:
          type: string
        isDefault:
          type: boolean

    # Settings Schemas
...